Transactions refactored
